Added a 2nd macro that delays restart if installed
macro CheckNetFrameworkDelayRestart
takes the same .NET version number as original version, but also returns if a .NET installer was run and also delays the restart of the .NET Installer.

The issue is if .NET is currently being use, and you do install a new version, the installer would automatically reboot the computer while in the provided macro.  This change allows the installer maintainer to control when the reboot happens, if it is needed.